added class Cart to Model, added two Enums, fixed DaoImpl

+++ src/Model/Utils/DAOs/BookDAO.java
@@ -1,6 +1,8 @@
 package Model.Utils.DAOs;
 
 import Model.Book;
+
+import java.math.BigDecimal;
 import java.sql.SQLException;
 
 public interface BookDAO {
@@ -11,7 +13,7 @@ public interface BookDAO {
 
     public boolean isAvailable(String ISBN) throws SQLException;
 
-    public void setDiscount(String ISBN, float discount) throws SQLException;
+    public void setDiscount(String ISBN, BigDecimal discount) throws SQLException;
 
     public void editDescription(String ISBN, String newDescription) throws SQLException;
 }
