+++ src/Model/Utils/DAOs/OrderDAO.java
@@ -3,6 +3,7 @@ package Model.Utils.DAOs;
 import Model.Order;
 import Model.User;
 
+import java.math.BigDecimal;
 import java.sql.SQLException;
 import java.util.ArrayList;
 
@@ -16,7 +17,7 @@ public interface OrderDAO {
     public ArrayList<Order> getOrdersByUser(String email) throws SQLException;
 
     public void addOrder(String orderID, String date, String status, String paymentMethod,
-                         float price, int points, String shippingAddress, String user,
+                         BigDecimal price, int points, String shippingAddress, String user,
                          String user_notReg) throws SQLException;
 
     public void updateStatus(Order order, String newStatus) throws SQLException;
